Put the Registry in the RepositoryManager
