+++ crates/core/src/lib.rs
@@ -3,5 +3,4 @@
 #![feature(stmt_expr_attributes)]
 
 pub mod change;
-pub mod registry;
 pub mod repository;
