--- crates/memory/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "solipr-memory"
-version = "0.0.1"
-description = "An implementation of in-memory data structures for Solipr"
-repository = "https://github.com/solipr/solipr"
-readme = "README.md"
-license = "GPL-3.0-or-later"
-keywords = ["vcs"]
-categories = ["development-tools"]
-edition = "2021"
-
-[dependencies]
-solipr-core = { path = "../core" }
-sha2 = "0.10.8"
-
-[lints]
-workspace = true
