--- crates/memory/src/lib.rs
@@ -1,4 +1,0 @@
-//! An implementation of in-memory data structures for Solipr that stores data
-//! in memory.
-
-pub mod registry;
