+++ crates/persistent/src/lib.rs
@@ -1,5 +1,4 @@
 //! An implementation of persistent data structures for Solipr that stores data
 //! on disk.
 
-pub mod registry;
 pub mod repository;
