+++ crates/persistent/src/repository.rs
@@ -2,19 +2,25 @@
 //! stores data in persistent storage (on disk).
 
 use std::collections::HashSet;
-use std::io;
-use std::path::Path;
+use std::fs::{self, File};
+use std::io::{self, Read, Write};
+use std::path::{Path, PathBuf};
 
+use base64::prelude::*;
 use fjall::{
     Config, Error, PartitionCreateOptions, ReadTransaction, Slice, TransactionalKeyspace,
     TransactionalPartitionHandle, WriteTransaction,
 };
+use sha2::{Digest, Sha256};
 use solipr_core::change::{Change, ChangeContent, ChangeHash, FileId, LineId, SingleId};
-use solipr_core::repository::{Repository, RepositoryId, RepositoryManager};
+use solipr_core::repository::{ContentHash, Repository, RepositoryId, RepositoryManager};
 
 /// An implementation of the [`RepositoryManager`] that stores data in
 /// persistent storage (on disk).
 pub struct PersistentRepositoryManager {
+    /// The folder where the data of the registry is stored.
+    registry_folder: PathBuf,
+
     /// The keyspace used to store data.
     ///
     /// It is an handle to the database.
@@ -48,13 +54,15 @@ impl PersistentRepositoryManager {
     ///
     /// An error will be returned if the folder could not be opened.
     pub fn create(folder: impl AsRef<Path>) -> Result<Self, Error> {
-        let keyspace = Config::new(folder).open_transactional()?;
+        let folder = folder.as_ref();
+        let keyspace = Config::new(folder.join("repositories")).open_transactional()?;
         let changes = keyspace.open_partition("changes", PartitionCreateOptions::default())?;
         let reverse_heads =
             keyspace.open_partition("reverse_heads", PartitionCreateOptions::default())?;
         let heads = keyspace.open_partition("heads", PartitionCreateOptions::default())?;
         let lines = keyspace.open_partition("lines", PartitionCreateOptions::default())?;
         Ok(Self {
+            registry_folder: folder.join("registry"),
             keyspace,
             changes,
             reverse_heads,
@@ -86,6 +94,74 @@ impl RepositoryManager for PersistentRepositoryManager {
             manager: self,
             transaction: RepositoryTransaction::Write(self.keyspace.write_tx()),
         })
+    }
+
+    fn read_content(&self, hash: ContentHash) -> Result<Option<impl Read>, Self::Error> {
+        let encoded_hash = BASE64_URL_SAFE_NO_PAD.encode(hash.as_bytes());
+        #[expect(
+            clippy::string_slice,
+            reason = "there is always more than 2 characters in a ContentHash"
+        )]
+        match File::open(
+            self.registry_folder
+                .join(&encoded_hash[..2])
+                .join(&encoded_hash[2..]),
+        ) {
+            Ok(file) => Ok(Some(file)),
+            Err(err) if err.kind() == io::ErrorKind::NotFound => Ok(None),
+            Err(err) => Err(err.into()),
+        }
+    }
+
+    fn write_content(&self, mut content: impl Read) -> Result<ContentHash, Self::Error> {
+        // Create the folder if it doesn't exist
+        if !self.registry_folder.exists() {
+            fs::create_dir_all(&self.registry_folder)?;
+        }
+
+        // Create a temporary file to store the content in
+        let temp_file_path = self.registry_folder.join(uuid::Uuid::now_v7().to_string());
+        let mut temp_file = File::create(&temp_file_path)?;
+
+        // Loop 32 bytes at a time and update the hasher
+        // until we reach the end of the content
+        let mut hasher = Sha256::new();
+        let mut buffer = [0; 32];
+        loop {
+            let byte_count = match content.read(&mut buffer) {
+                Ok(0) => break,
+                Ok(byte_count) => byte_count,
+                Err(ref err) if err.kind() == io::ErrorKind::Interrupted => continue,
+                Err(err) => return Err(err.into()),
+            };
+            #[expect(
+                clippy::indexing_slicing,
+                reason = "byte_count is always smaller or equal than 32"
+            )]
+            hasher.update(&buffer[..byte_count]);
+            #[expect(
+                clippy::indexing_slicing,
+                reason = "byte_count is always smaller or equal than 32"
+            )]
+            temp_file.write_all(&buffer[..byte_count])?;
+        }
+        temp_file.flush()?;
+        drop(temp_file);
+
+        // Create a unique hash for the content
+        let hash = ContentHash::new(hasher.finalize().into());
+        let encoded_hash = BASE64_URL_SAFE_NO_PAD.encode(hash.as_bytes());
+
+        // Move the temporary file into the correct location
+        let (subfolder, file) = encoded_hash.split_at(2);
+        let path_dir = self.registry_folder.join(subfolder);
+        if !path_dir.exists() {
+            fs::create_dir_all(&path_dir)?;
+        }
+        fs::rename(temp_file_path, path_dir.join(file))?;
+
+        // Return the hash of the content
+        Ok(hash)
     }
 }
 
@@ -133,6 +209,12 @@ impl PersistentRepository<'_> {
 
 impl<'manager> Repository<'manager> for PersistentRepository<'manager> {
     type Error = Error;
+
+    type Manager = PersistentRepositoryManager;
+
+    fn manager(&self) -> &'manager Self::Manager {
+        self.manager
+    }
 
     fn changes(&self) -> impl Iterator<Item = Result<(ChangeHash, Change), Self::Error>> {
         let iter: Box<dyn Iterator<Item = Result<(Slice, Slice), _>>> = match self.transaction {
