+++ crates/tests/Cargo.toml
@@ -12,6 +12,5 @@ edition = "2021"
 [dev-dependencies]
 solipr-core = { path = "../core" }
 solipr-persistent = { path = "../persistent" }
-solipr-memory = { path = "../memory" }
 tempfile = "3.13.0"
 rand = "0.8.5"
