+++ crates/tests/src/lib.rs
@@ -2,4 +2,4 @@
 
 #![cfg(test)]
 
-mod registry;
+mod repository;
